Worked on Sockets
Currently having issues with gevent-socketio, in particular the
SocketIOServer. According to the stacktrace there is a syntax error on
ln 108, but I doubt this very much. (Worth noting I lack the experience
with Python to know for sure whether this is my bad or that is
legitimately a syntax error).