Final Build
These are some final changes I made before handing the project in. It is
working in its current state, however is incredibly simple. I do not
wish to build this further, due to my lack of confidence in some of the
libraries I made use of throughout this project. Moving on to something
bigger and better now!